# api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: imajin events
  version: "1.0.0"
  description: >
    Event creation, ticketing, ticket tiers, queue management, hold system, .fair manifests, checkout, and payment webhooks.
    Served by the app itself at /api/spec. Authenticate with a scoped app token (`Authorization: Bearer`, audience `events`,
    scopes `events:read` for GET/HEAD/OPTIONS and `events:write` otherwise) or, during migration, the legacy session cookie.
servers:
  - url: https://events.imajin.ai
    description: production
  - url: https://dev-events.imajin.ai
    description: dev
  - url: http://localhost:3000
    description: local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: imajin_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Scoped app token minted for audience `events` (events:read for safe methods, events:write otherwise)"
    webhookSecret:
      type: http
      scheme: bearer
      description: "Bearer WEBHOOK_SECRET — internal payment webhook from pay service"
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ticket not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
paths:
...
  /api/attending/{did}:
    get:
      tags: [events]
      operationId: listAttending
      summary: "Upcoming events a DID holds a ticket for (profile page)"
      description: "Public. A ticket counts as held when its status is valid, sold or used (#2734). invite_only events are only shown if viewer_did also holds a ticket."
      parameters:
        - { name: did, in: path, required: true, schema: { type: string } }
        - { name: viewer_did, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: Upcoming events; invite-only events are listed only when viewer_did also holds a ticket, content: { application/json: { schema: { type: object } } } }
        "500": { description: Failed to load attending events, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /api/campaign/{eventId}/my-pledge:
    get:
      tags: [events]
      operationId: getMyPledge
      summary: "The caller's pledge for a campaign, if any"
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: The caller's pledge (or null), content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { description: Failed to load pledge, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/campaign/{eventId}/pledges:
    get:
      tags: [events]
      operationId: listCampaignPledges
      summary: "All pledges for a campaign (creator only)"
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Pledges, content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { description: Failed to load pledges, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/campaign/{eventId}/cancel:
    post:
      tags: [events]
      operationId: cancelCampaign
      summary: "Cancel a campaign and all pending/confirmed pledges (creator only)"
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Number of pledges cancelled, content: { application/json: { schema: { type: object, properties: { cancelled: { type: integer } } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { description: Rate limited, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "500": { description: Failed to cancel campaign, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/campaign/{eventId}/settle:
    post:
      tags: [events]
      operationId: settleCampaign
      summary: "Charge all confirmed pledges (creator only)"
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Settlement results, content: { application/json: { schema: { type: object, properties: { charged: { type: integer }, failed: { type: integer }, total: { type: integer }, results: { type: array, items: { type: object, properties: { pledgeId: { type: string }, status: { type: string }, error: { type: string } } } } } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { description: Rate limited, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "500": { description: Settlement failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /api/events/{id}/tickets/{ticketId}/cancel:
    post:
      tags: [events]
      operationId: cancelTicket
      summary: "Cancel a held or available (unconfirmed) ticket"
      description: "Only tickets in status held or available; confirmed tickets must use refund."
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: ticketId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Ticket cancelled, content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...